+++ page_double.go
@@ -1,6 +1,9 @@
 package parquet
 
 import (
+	"io"
+
+	"github.com/parquet-go/bitpack/unsafecast"
 	"github.com/parquet-go/parquet-go/encoding"
 )
 
@@ -67,4 +70,35 @@ func (page *doublePage) makeValue(v float64) Value {
 	value := makeValueDouble(v)
 	value.columnIndex = page.columnIndex
 	return value
+}
+
+type doublePageValues struct {
+	page   *doublePage
+	offset int
+}
+
+func (r *doublePageValues) Read(b []byte) (n int, err error) {
+	n, err = r.ReadDoubles(unsafecast.Slice[float64](b))
+	return 8 * n, err
+}
+
+func (r *doublePageValues) ReadDoubles(values []float64) (n int, err error) {
+	n = copy(values, r.page.values[r.offset:])
+	r.offset += n
+	if r.offset == len(r.page.values) {
+		err = io.EOF
+	}
+	return n, err
+}
+
+func (r *doublePageValues) ReadValues(values []Value) (n int, err error) {
+	for n < len(values) && r.offset < len(r.page.values) {
+		values[n] = r.page.makeValue(r.page.values[r.offset])
+		r.offset++
+		n++
+	}
+	if r.offset == len(r.page.values) {
+		err = io.EOF
+	}
+	return n, err
 }
