+++ page_repeated.go
@@ -1,6 +1,8 @@
 package parquet
 
 import (
+	"io"
+
 	"github.com/parquet-go/parquet-go/encoding"
 )
 
@@ -109,4 +111,62 @@ func (page *repeatedPage) Slice(i, j int64) Page {
 		repetitionLevels[rowIndex1:rowIndex2:rowIndex2],
 		definitionLevels[rowIndex1:rowIndex2:rowIndex2],
 	)
+}
+
+type repeatedPageValues struct {
+	page   *repeatedPage
+	values ValueReader
+	offset int
+}
+
+func (r *repeatedPageValues) ReadValues(values []Value) (n int, err error) {
+	maxDefinitionLevel := r.page.maxDefinitionLevel
+	definitionLevels := r.page.definitionLevels
+	repetitionLevels := r.page.repetitionLevels
+	columnIndex := ^int16(r.page.Column())
+
+	// While we haven't exceeded the output buffer and we haven't exceeded the page size.
+	for n < len(values) && r.offset < len(definitionLevels) {
+
+		// While we haven't exceeded the output buffer and we haven't exceeded the
+		// page size AND the current element's definitionLevel is not the
+		// maxDefinitionLevel (this is a null value), Create the zero values to be
+		// returned in this run.
+		for n < len(values) && r.offset < len(definitionLevels) && definitionLevels[r.offset] != maxDefinitionLevel {
+			values[n] = Value{
+				repetitionLevel: repetitionLevels[r.offset],
+				definitionLevel: definitionLevels[r.offset],
+				columnIndex:     columnIndex,
+			}
+			r.offset++
+			n++
+		}
+
+		i := n
+		j := r.offset
+		// Get the length of the run of non-zero values to be copied.
+		for i < len(values) && j < len(definitionLevels) && definitionLevels[j] == maxDefinitionLevel {
+			i++
+			j++
+		}
+
+		// Copy all the non-zero values in this run.
+		if n < i {
+			for j, err = r.values.ReadValues(values[n:i]); j > 0; j-- {
+				values[n].repetitionLevel = repetitionLevels[r.offset]
+				values[n].definitionLevel = maxDefinitionLevel
+				r.offset++
+				n++
+			}
+			if err != nil && err != io.EOF {
+				return n, err
+			}
+			err = nil
+		}
+	}
+
+	if r.offset == len(definitionLevels) {
+		err = io.EOF
+	}
+	return n, err
 }
